Refactor shaders into it's own directory.
